refactor(team): memeber 엔티티를 Memeber 도메인의 엔티티로 교체
