+++ Backend/src/main/java/com/luckyseven/backend/domain/team/controller/TeamController.java
@@ -1,12 +1,13 @@
 package com.luckyseven.backend.domain.team.controller;
 
+import com.luckyseven.backend.domain.member.entity.Member;
 import com.luckyseven.backend.domain.team.dto.TeamCreateRequest;
 import com.luckyseven.backend.domain.team.dto.TeamCreateResponse;
 import com.luckyseven.backend.domain.team.dto.TeamDashboardResponse;
 import com.luckyseven.backend.domain.team.dto.TeamJoinRequest;
 import com.luckyseven.backend.domain.team.dto.TeamJoinResponse;
 import com.luckyseven.backend.domain.team.dto.TeamMemberDto;
-import com.luckyseven.backend.domain.team.entity.Member;
+
 import com.luckyseven.backend.domain.team.service.TeamMemberService;
 import com.luckyseven.backend.domain.team.service.TeamService;
 import com.luckyseven.backend.domain.team.service.TempMemberService;
