+++ Backend/src/main/java/com/luckyseven/backend/domain/team/entity/Expense.java
@@ -1,8 +1,12 @@
 package com.luckyseven.backend.domain.team.entity;
 
+import com.luckyseven.backend.domain.member.entity.Member;
 import com.luckyseven.backend.sharedkernel.entity.BaseEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
@@ -24,6 +28,9 @@ import lombok.NoArgsConstructor;
 })
 
 public class Expense extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
     
     @ManyToOne
     @JoinColumn(name = "team_id")
