+++ Backend/src/main/java/com/luckyseven/backend/domain/team/service/TempMemberService.java
@@ -1,6 +1,6 @@
 package com.luckyseven.backend.domain.team.service;
 
-import com.luckyseven.backend.domain.team.entity.Member;
+import com.luckyseven.backend.domain.member.entity.Member;
 import com.luckyseven.backend.domain.team.repository.TempMemberRepository;
 import com.luckyseven.backend.sharedkernel.exception.CustomLogicException;
 import com.luckyseven.backend.sharedkernel.exception.ExceptionCode;
