+++ Backend/src/main/java/com/luckyseven/backend/sharedkernel/entity/BaseEntity.java
@@ -23,10 +23,6 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 @NoArgsConstructor
 @EntityListeners(AuditingEntityListener.class)
 public abstract class BaseEntity {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
   
   @CreatedDate
   @Column(updatable = false, nullable = false)
